+++ tests/test_app/wsgi.py
@@ -1,3 +1,6 @@
-from .routes import api
+try:
+    from .routes import api
+except ImportError:
+    from routes import api
 
 app = api